before i do anything...

+++ TheForest/TheForest/AudioManager.h
@@ -46,7 +46,7 @@ private:
     Mix_Music* soundPlayer;
     std::vector<Mix_Chunk*> sounds;
 
-    static constexpr int soundsLength = 13;
+    static constexpr short soundsLength = 13;
     
     // The order of the array affects the selection of the sound
     const std::string soundPaths[soundsLength]
