+++ TheForest/TheForest/FlowerManager.cpp
@@ -74,12 +74,10 @@ void FlowerManager::Level1Flowers()
     const FlowerUp one = FlowerUp(rPlayer, {1750, 228});
     const FlowerUp two = FlowerUp(rPlayer, Vector2(GameWindow::GetWindowWidth() * 2 + 900, 450));
     const FlowerUp three = FlowerUp(rPlayer, Vector2( GameWindow::GetWindowWidth() * 4 + 1550, 520));
-    // FlowerUp four = FlowerUp(rPlayer, Vector2( GameWindow::GetWindowWidth() * 5 + 420, 970));
 
     flowers.push_back(one);
     flowers.push_back(two);
     flowers.push_back(three);
-    // flowers.push_back(four);
 }
 
 void FlowerManager::Level2Flowers()
@@ -134,7 +132,7 @@ FlowerManager::FlowerUp::FlowerUp(Player& player, Vector2 pos): position(pos), r
 {
     int screenWidth;
     SDL_GetWindowSize(GameWindow::GetWindow(), &screenWidth, nullptr);
-    levelSlide = floor(position.x / screenWidth);
+    levelSlide = static_cast<short>(floor(position.x / screenWidth));
 }
 
 void FlowerManager::FlowerUp::Update(float deltaTime)
